refactor(hfb): move HFB RFI flagging to subpackage
